bpf-verification: Refactor to prepare for parallelization
This commit moves all per-instruction logic for gen verification into
its own function with clear inputs and outputs. The goal is to have the
big code movements here, to make the actual paralellization trivial in
the next commit.

This commit was tested by running the verification for Linux v5.9 and
checking that the results didn't change.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>