bpf-verification: Parallelize gen verification
This commit uses the multiprocessing Python module to parallelize the
execution of the gen verification on as many cores as the machine has.
The gen verification was selected because it currently takes most of the
execution time.

Parallelism is achieved by running the solver for each instruction on
different cores. By default, the multiprocessing Pool will spawn as many
processes as there are cores. Those processes will then execute the work
items in turn until they are all done.

Input data (pool.apply_async's second argument) is copied into the
processes. Changes to those variables is therefore not reflected in the
caller or in other processes.

We also need to change the stdout text. "10/15 Verifying BPF_SUB ...
Done." doesn't work anymore because it's not all printed at once.
Instead, we print one line for the start of the job and one for the end:

    1/2 Verifying BPF_XOR_32...
    2/2 Verifying BPF_ADD...
    1/2 BPF_XOR_32 Done.

This commit was tested by running the verification for v5.9 and checking
that the results are as before.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>