bpf-verification: Remove bugs_dict parameter
bug_dict is a dictionary that keeps track of BPF instruction and bound
violation types u64, s64... pair. It looks like:

    {"bpf_or": ["u64","s64","tnum","u32","s32"], ...}

It was needed to keep track of which BPF instruction is violating which
bound and then synthesize a program accordingly. At the start of the
verification, each BPF instruction has 5 possible violations which get
diminished as the gen and sro verifications take place. Whatever bound
violation is left when synthesis stage is reached will be synthesized.

In the config file, bug_dict was used in case you wanted to only
explore a specific bug type of some bpf instruction. For example, let's
say you had a hunch that bpf_or is violating only u64 so you'd only try
to verify that specific bug type for that specific instruction. That
config doesn't seem needed at this point anymore since it probably
won't save much time.

Co-authored-by: Matan Shachnai <m.shachnai@rutgers.edu>
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>