bpf-verification: Remove TOML config
This commit removes the TOML config file and all related logic. Having
two ways to provide parameters complicates the logic and the
instructions. Those two inputs also tend to diverge over time (cf.
bugs_dict which was only available in the TOML config).

Between the two, the inline parameters method is prefered because it's
easier to work with in CI and in the command line (no need to edit a
file). It's also the de facto standard for Python scripts and allows us
to set some constraints on parameters.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>