llvm-to-smt: Support LLVM IR switch instructions
Since Linux commit 904e6ddf4133 ("bpf: Use scalar ids in
mark_chain_precision()") we may need to handle a switch in the LLVM IR.

    if.end.i.i:                                       ; preds = %if.then.i
      %i.i.i.i = and i32 %i.i.i, 255
      %trunc = trunc i32 %i.i.i to i8
      switch i8 %trunc, label %lor.rhs.i.i [
        i8 13, label %reg_not_null.exit.i
        i8 12, label %reg_not_null.exit.i
        i8 11, label %reg_not_null.exit.i
        i8 5, label %reg_not_null.exit.i
        i8 4, label %reg_not_null.exit.i
        i8 16, label %land.lhs.true.i.i
      ]

Because we don't currently handle switches, the conversion to SMT fails
with "<incomingBlock, currentBB> pair not found in EdgeAssertionsMap".
The EdgeAssertionsMap map should hold the Z3 expression leading to each
destination basic block of the switch. It doesn't because we only
populate this map for branch instructions.

We therefore need the equivalent of handleBranchInst but for
SwitchInst. This commit adds such a function, to populate both
EdgeAssertionsMap and pathConditionsMap.

Instead of converting each switch case into a EdgeAssertionsMap entry,
we recognize that several cases may lead to the same destination basic
block. We therefore conjunct such cases. To that end, a first pass over
all cases build a map (destination basic block => Z3 expr). This map is
then used to build both EdgeAssertionsMap and pathConditionsMap.

It's unclear if conjuncting the cases like that actually lowers the Z3
runtime, but I figure it can't hurt.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>