llvm-to-smt: Fix wrong 64-bit wrapper for 5.1 and 5.2
It is not enough to introduce the 32-bit variant of wrapper_jmp for
upstream commit v5.1-rc1~178^2~404^2~4^2~13. That commit also changed
the 64-bit variant so we need a new wrapper in between our version 2 and
3.

This commit shifts all variants above 2 by 1 to introduce a new
wrapper_jmp_3 for kernel versions v5.1-rc1 to v5.3-rc1 (unincluded).

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>