Update handling phi instructions and handle loads derived from phi
1. Split the logic of handlePhiNode into handlePhiNodeSetupBitVecs and handlePhiNodeResolvePathConditions:
- handlePhiNodeSetupBitVecs is called in pass #1.
It sets up the path condition booleans and saves them in phiResolutionMap. Additionally, it
stores the entire phi instruction's structure in phiMap (to be used later).
- handlePhiNodeResolvePathConditions is called in pass #3, when the path conditions for each incoming
edge to the basic block has been figured out. The path condition booleans created in
handlePhiNodeSetupBitVecs are set appropritately.

2. Handle loads derived from phi instructions
- Handle this in a similar way to handling loads derived from select instructions.
When encountering such a load use the information populated in PhiMap and GEPMap to
resolve the encoding(s) for the load.