llvm-to-smt: Update wrappers for 6.8-rc1+
Starting from 6.8-rc1, a new suffix function of abstract operators was
introduced, reg_bounds_sanity_check, which is called after reg_bounds_sync.
If we want to perform modular verification, we need to account for
this function. This commit injects the logic of reg_bounds_sanity_check
into reg_bounds_sync before obtaining the encoding for reg_bounds_sync.
This way an op of the form (op.sync.sanity) is encoded modularly as
(op), (sync.sanity).