workflows: CI for the verification step
This commit adds a GitHub CI test for bpf_alu_jmp_synthesis.py and
related logic. It is currently limited to the v5.9 kernel as newer
kernels easily take hours to verify a single instruction. It's also
limited to two instructions (one jump, one ALU) among the fatest to
verify. Total verification time appears to be <5min.

This workflow will run for pull request (when opened or updated) and for
all pushes to main. If several commits are pushed to main at the same
time, it should only run for the latest commit of the batch.

There is a lot that can be improved: running this only when
bpf-verification/ is modified, covering more kernels and instructions,
checking the results, etc. I chose to start with something small we can
improve later on.

Coverage for llvm-to-smt is still a work in progress as I'm hitting
weird some memory consumption behavior on GitHub.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>