workflows: Initial CI for the LLVM-to-SMT step
This commit adds a GitHub CI workflow for generate_encodings.py and
associated logic. Total runtime is below three minutes.

This workflow will run for pull request (when opened or updated) and for
all pushes to main. If several commits are pushed to main at the same
time, it should only run for the latest commit of the batch.

It currently doesn't support running against latest as that will require
updating the LLVM dependency. The resulting encodings are also not
checked yet. It simply verifies that the script doesn't error out.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>