llvm-to-smt: Use LLVM 14.0.0
Since Linux commit 9c1b86f8ce04e ("kbuild: raise the minimum supported
version of LLVM to 13.0.1"), which is v6.9-rc1~106^2~38, the kernel
requires LLVM 13.0.1 to compile.

We therefore need to upgrade our LLVM dependency to be able to run Agni
against latest kernels. Upgrading past LLVM 15 is a lot more work
because typed pointers were replaced by opaque pointers, but we should
be able to upgrade to LLVM 14.

Two changes are needed in the sources to adapt to changes in the LLVM
API:
- sys::fs::F_None was renamed to sys::fs::OF_None.
- The handy Builder.CreateLoad functions that take only two arguments
  have been removed.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>