llvm-to-smt: Revert back to using O1 passes
run_llvm_passes.py was updated in d52bff5 to use Oz passes,
(that optimize for code size) instead of O1 passes. This worked in
llvm-12, but not anymore when updating to llvm-14. Ostensibly,
the passes bundled in "Oz" have changed in llvm-14. Revert back
to using O1. This potentially increases code size, and smt2 encoding
size.