generate_encodings: Remove memset from __mark_reg_unknown_imprecise
Older kernels made a call to memset in "__mark_reg_unknown"
in verifier.c. As a pre-processing step, the function
mark_reg_unknown_memset_remove() deleted this function call
so that LLVMToSMT does not have to deal with "llvm.memset"
intrinsics.

Newer kernels (~6.9) renamed this function to
"__mark_reg_unknown_imprecise".

This commit, attempts to remove the call to "memset" in
__mark_reg_unknown_imprecise first. If it is unable to remove it,
then it attempts to remove the call to "memset" in
mark_reg_unknown.

Co-authored-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>