workflows: Compare encodings using git
This comparison happens in two steps. Before generating new encodings,
we prepare a git repository containing the existing encodings, with .ll
and .log files not committed (i.e., only .smt2).

Newly-generated encodings are then added to this repository. After
generation, git status is used to print the list of files that changed
if any. If any file changed, the job is failed.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>