bpf-encodings: Replace encodings with CI encodings for LTS kernels
This commit removes all the previously-generated encodings and replaces
them with encodings generated by the GitHub Action CI for LTS kernels.
Kernel 6.8 is also added because a lot changed in that kernel and there
isn't yet a newer LTS.

Keeping encodings for every single kernel in git is probably not viable
in the long term. Limiting it to LTS kernels allows us to focus on the
kernels who matter.

Using the encodings generated by the GitHub CI will allow us to compare
the future CI results against this known good version.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>