generate_encodings: Fix fls wrappers
fls and fls64() are functions used in tnum.c. They are externally
defined and their definition is platform dependent, usually
using asm instructions native to the platform. We provide our
own platform agnostic definitions using kernel sources[1], so
that llvm-to-smt can generate encodings without having to
handle special llvm intrinsics like llvm.ctlz.

However, the definition of fls that was being used previously,
was producing incorrect results.
Previously,
s32 a = 0xff0ff0ff;
fls(a) => 63, which is incorrect.

The definition of fls also needs to depend on the BITS_PER_LONG
on the platform. In this commit, we provide both definitions, one
to be used when BITS_PER_LONG is 32, and one for 64.

After the fix, fls(a) => 32 the correct answer

[1] https://elixir.bootlin.com/linux/v6.10/source/include/asm-generic/bitops/fls.h

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>