workflows: End-to-end test
This new workflow will run on a schedule to validate the linus and
bpf-next trees. It will both generate the encodings for those trees and
run the verification, on a couple instructions.

This workflow is split into two sets of jobs. First two jobs download
the Linux sources. Second, one job per instruction and per tree (for a
total of 66 jobs currently) generates the encodings and runs
verification.

If we try to download the Linux sources directly in the second set of
jobs, we end up with up to 66 parallel downloads. That gets throttled by
GitHub to a painfully slow download speed and often ends up just
failing. It's a lot faster to download each tree only once, compress it
and pass it along to the subsequent jobs.

We need to compress the trees before uploading/downloading them because
the actions/upload-artifact GitHub action isn't meant to upload such a
large set of files. If we tried to upload the whole sources without
compressing, it tries to open too many file descriptors at once and
fails.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>