llvm-to-smt: fix wrappers for 6.8-rc1
This commit fixes the wrappers for 6.8-rc1 to include
the "fake_reg" part of the verifier logic (that ends
up using the same src_reg for exploring both branches,
when the instruction is a BPF_JMP_IMM. This lead
to a CVE [1], which Agni didn't detect because
we don't support BPF_JMP_IMM instructions yet.

Currently, we only support BPF_JMP_REG instructions,
so this commit does not affect anything directly.
In the future though when we support BPF_JMP_IMM,
this update should allow us to catch the fake_reg issue.

This update also uses non-C-style comments for
commenting on "our code" in order differentiate from
kernel code that uses C-style comments.

[1]: https://github.com/bpfverif/agni/issues/40

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>