llvm-to-smt: Add fls wrappers for verifier.c
Incoming commit [1] to verifier.c changes scalar*_min_max_and, by
calling a function negative(32)_bit_floor, which uses the fls(64)
functions.

The calls in verifier.c to fls(64) need to be changed to use our
platform agnostic versions fls___, and fls64___ that we provide
in wrappers.py borrowed from kernel sources [2]. This commit replaces
all the calls to fls and fls64 *in verifier.c* to the ones that we
provide.

[1]: https://lore.kernel.org/bpf/20240719110059.797546-6-xukuohai@huaweicloud.com/
[2]: https://elixir.bootlin.com/linux/v6.10/source/include/asm-generic/bitops/fls.h

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>