workflows/end-to-end: Optimize artifact uploads
We upload artifacts for the encoding generation and the BPF
verification, regardless of whether they fail or not. There's however
no point in uploading those artifacts if the corresponding steps didn't
even run. So for example, if the encoding generation fails, we will
upload encoding artifacts, but we won't upload BPF verification
artifacts since we never got to that step.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>