llvm-to-smt: introduce unique bitvector suffixes and bpf_op_attrs
For every BPF_* instruction that we want to obtain encodings for,
generate_encodings.py adds a unique suffix to every bitvector in the
formula, when it calls LLVMPassRunner [1].

It is useful to set the bitvector suffixes statically in code,
so that a particular op always uses the same suffix across multiple
runs. This will be true even when --specific-op is used.
See also [2].

This commit introduces a separate class called bpf_op_attrs to
manage the attributes of all instructions that we want to encode.
The final list of ops that we want to obtain encodings for is
a list of bpf_op_attrs called "bpf_ops". When creating this list
we assign a hardcoded "suffix_id" field for every op. This
id should never change, even across commits. New ops should
be added to this list with new hardcoded suffix_ids.

It is also useful to track the some properties of a bpf_op
together, hence the bpf_op_attrs class was introduced, instead
of using first class python lists. bpf_op_attrs has the following
properties:
- "skip", to be used to decide which op generation should be skipped
- "intro_ver", when this op was introduced, insted of just comments
- "insn_class", can also be used hen we introduce
support BPF_JMP_IMM, BPF_JMP32_IMM, BPF_ALU32_IMM, BPF_ALU64_IMM

[1]: https://github.com/bpfverif/agni/blob/8e62e73a50b92c420a89031f362c3dcfdee617e5/llvm-to-smt/generate_encodings.py#L867
[2]: https://github.com/bpfverif/agni/issues/48