workflows/llvm-to-smt: Remove moving kernel targets
We added coverage for the bpf-next and linus trees in the llvm-to-smt
workflow when we didn't have the end-to-end workflow. Back then it was
the only way to cover those trees, but it's unideal. These "kernel
targets" are not static and changes in the kernel may introduce
regressions in Agni. By running those as part of the pull request CI, we
risk to block pull requests on upstream issues.

Instead, we are now covering those kernels in the schedule, end-to-end
workflow. We can therefore remove them from the llvm-to-smt workflow.
This commit adds the latest released kernel, v6.10, instead.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>