llvm-to-smt: Support llvm.*.with.overflow.* intrinsics and extractvalue
This commits support for llvm add/sub intrinsics with overflow
detection: llvm.sadd.with.overflow.*, llvm.uadd.with.overflow.*,
llvm.ssub.with.overflow.*, llvm.usub.with.overflow.* . It also adds
limited support for extractvalue instructions to support the intrinsics.
This is needed in order to support BPF_ADD and BPF_SUB instructions in
v6.10 [1].

The return type of the add with overflow intrinsic is an aggregate type.
The types are automatically inferred and appropriate sized bitvectors
are used (i32/i64/any other size) in the formula. This commit also adds
support for extractvalue instructions, that typically occur *after* the
intrinsic call instructions.

Example:
%res = call {i32, i1} @llvm.sadd.with.overflow.i32(i32 %a, i32 %b)

This instruction performs a signed addition of the two variables, and
returns a structure — the first element of which is the signed
summation, and the second element of which is a bit specifying if the
signed summation resulted in an overflow.

Consider the above instruction. We first create two bitvectors and store
them in a bitvector tree: [res_i32_bv, res_i1_bv]. We store this tree in
the global map BBValueBVTreeMap to be retreived later. Finally, we
assert that: if the signed addition did *not* overflow, then the
bitvector res_i1_bv (that corresponds to the whether the result
*overflowed* in llvm) is set to 0, else it is set to 1.

According to llvm LangRef [2] the semantics of llvm.sadd.with.overflow
is defined such that it is meant to detect both positive overflow (sum
of two positive numbers yields a negative result) and negative overflow
(sum of two negative numbers yields a positive result). The semantics is
similarly defined for other llvm.*.with.overflow intrinsics.

For addition, Z3 has the predicates BVAddNoOverflow, BVAddNoUnderflow.
Per Z3, an overflow occurs when there is positive overflow and underflow
occurs when there is negative overflow.

So, in order to encode the semantics of llvm.sadd.with.overflow, we need
to assert in Z3 that either overflow or underflow occured:
(!BVAddNoOverflow \/ !BVAddNoUnderflow). Alternatively, (BVAddNoOverflow
/\ BVAddNoUnderflow) asserts overflow did *not* occur.

z3.If(
  z3.And(z3.BVAddNoOverflow(a_bv, b_bv, True), z3.BVAddNoUnderflow(a_bv, b_bv)),
  res_i1_bv == 0,
  res_i1_bv == 1
)

res_i32_bv == a_bv + b_bv

[1]: https://github.com/bpfverif/agni/issues/44
[2]: https://llvm.org/docs/LangRef.html#arithmetic-with-overflow-intrinsics

Suggested-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>