workflows: Cancel previous workflow on push
The goal of this commit is to avoid (or rather cancel) useless workflow
runs.

The two workflows modified are triggered whenever someone updates a
pull request. If someone pushes to the same pull request twice in a
short time, the first workflow runs may not be finished before the
second push. That is, old workflow runs keep running even if the code
they are testing/running has been overwritten.

Concurrency groups allow us to fix that. We can configure the
concurrency group such that only one job in the group runs at a given
time; when a new job is pushed, previous ones are cancelled. The
concurrency group is then identified by its name.

This commit defines two types of concurrency group names, depending on
how the workflow is triggered. If triggered on a pull request update:

    [workflow name]-[pull request number]

If triggered on a push to the main branch:

    [workflow name]-[SHA of last commit on the branch]

For example for the llvm-to-smt workflow, this ensures there will be
only one run of this workflow per pull request (resp. per push to the
main branch) at a given time.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>