workflows: Define timeouts
This commit defines timeouts as a safety measure, to avoid wasting too
many CI resources if something takes too long or never finishes.

To define those timeouts, I took the longest runs over the past week
(5min and 19min), rounded it up, and doubled it (20min and 40min). The
goal is not to be close to actual runtimes but to avoid wasting
resources in error cases, so its fine if we are well above.

For the end-to-end workflow, which is a lot longer (~2h), I rounded up
to 3h instead of rounding up and doubling.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>