llvm-to-smt: Address all instructions in populateBBAssertionsMap
BBAssertionsMapPass (i.e., the first pass over the llvm IR)
performed by function populateBBAssertionsMap(), is the only
pass that iterates over *all the instructions( in IR. Previously,
if an instruction was unknown or not supported, this function
just "continued" silently. It is important that we detect
instructions that are not supported yet.

This commit adds cases for handleBranchInst, and handleReturnInst
to populateBBAssertionsMap(). These cases essentially do nothing
in the BBAssertionsMapPass. However, we now have
complete coverage for all *instructions* in one function, i.e.
populateBBAssertionsMap(), which will now throw an exception
if a new instruction is encountered.

This is more of a nice-to-have for the future, and it doesn't
affect any encoding generation.

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>