generate-encodings: Add option for overriding bitvector suffixes
In order to compare two encodings for equivalent input-output behavior,
we need the pairs of encodings to use different bitvectors. Currently,
bitvector suffixes are hardcoded in the suffix_id field of each op in
bpf_ops. This commit allows overriding the default bitvector suffixes by
providing a flag that replaces the hardcoded_suffix_id with
hardcoded_suffix_id + 1000.

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>