workflows: Use Ubuntu 22.04
In the next commit, we need to install clang-12, which isn't available
on Ubuntu 24.04. So let's downgrade back to Ubuntu 22.04 for the time
being.

As a unexpected side effect, this commit also fixes the build for
kernels v6.1+. On Ubuntu 22.04, the clang-14 package is an older patch
version (i.e., 14.0.0) and doesn't have the --instnamer error from the
latest clang-14 package found on Ubuntu 24.04.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>