workflows: Use LLVM 16 in CI
The easiest way to install LLVM 16 is to switch to Ubuntu 24.04 and
install via the packages. This commit implements that change for the BPF
verification and the end-to-end workflows.

For the LLVM-to-SMT workflow however, we need to keep the glibc version
from Ubuntu 22.04 to compile older kernels. So instead, we'll use the
apt.llvm.org/llvm.sh script to install LLVM 16.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>