llvm-to-smt: Support llvm umin/umax/smin/smax intrinsics
After starting to use clang-16 to compile the kernel, we observe new
intrinsics emitted, for example:

%x = call i32 @llvm.smax.i32(i32 %a, i32 %b)

which sets x to the signed maximum of a and b.

To encode this we use z3::ite and z3::sge:
x_bv == z3::ite(
            z3::sge(a_bv, b_bv),
                    a_bv,
                    b_bv
        )
The other intrinsics are similarly encoded using z3::sle, z3::uge,
z3::ult.

Suggested-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>