bpf-encodings: Update encodings for new Clang
Because we'll update Clang in CI, the encodings will change. Given we
use those in the CI to they only change when they should, we need to
update the encodings we have even if they are functionally identical.

BPF_{JEQ,JNE}{,_32} encodings are missing for v6.6 due to errors. This
isn't blocking for tests as we only test BPF_AND and BPF_JSLT for v6.6.

This commit also replaces the encodings for 6.10 by those for LTS
version 6.12.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>