llvm-to-smt: Avoid duplicate SMT files
SMT files are currently copied after generation from the instruction
directory to the parent directory. Since we commit those files for LTS
kernels (to check changes in CI), this results in a very large commit
history. In worst cases, it can lead GitHub to fail to load pull
requests with HTTP 500 errors.

This commit avoid the duplication by instead moving the SMT files to
the parent directory instead of copying them.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>