llvm-to-smt: Support llvm ctpop intrinsics
In order to support BPF_JSET, we will have to add support for the
llvm.ctpop intrinsic, which is does a population count of 1 bits in an
integer, for example:

%o = call i4 @llvm.ctpop.i4(i4 %i)

counts the number of 1 bits in an input integer %i.

To encode in z3, we extract each individual bit of the input to a
1-width bitvector, and accumulate their sum in popCountSum. So, for a
4-bit integer argument %i (corresponding bitvector i_bv), we compute
popCountSum as:

ZeroExt(3, Extract(0, 0, i_bv))
+ ZeroExt(3, Extract(1, 1, i_bv))
+ ZeroExt(3, Extract(2, 2, i_bv))
+ ZeroExt(3, Extract(3, 3, i_bv))

Finally, we assert the the bitvector corresponding to output %o (i.e.
o_bv) is equal to the popCountSum:

o_bv == popCountSum_bv

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>