Verify reg_bounds_sync faster with further divide-and-conquer
The SAS'24 paper [1] explains how a divide-and-conquer approach allowed
Agni to verify the Linux range analysis faster. In particular,
reg_bounds_sync was verified separately from per-operation logic. This
change kept the verification time to minutes for a while, but in Linux
v6.8, it grew again to more than an hour and to several hours in Linux
v6.11. This is simply happening because reg_bounds_sync is becoming
smarter [2, 3, 4].

At Linux Plumbers 2024, we discussed [5] pushing the divide-and-conquer
approach further, by verifying reg_bounds_sync subfunctions separately.
This commit implements that idea.

The three subfunctions are __update_reg_bounds, __reg_deduce_bounds,
and __reg_bound_offset. They are mapped to pseudo instructions
BPF_SYNC1, BPF_SYNC2, and BPF_SYNC3. Most of the logic applied to
BPF_SYNC can then apply to these three instructions as well.

This ends up reducing the runtime from more than four hours to less than
ten minutes [6]. The downside of course is that verification will now
fail if those three functions are not independently sound.

BPF_SYNC is kept in the end-to-end CI for now, but can probably be
removed in the future, once we're sure this patch is working as
expected.

1 - https://harishankarv.github.io/assets/files/agni_sas24.pdf
2 - https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d7f008738171
3 - https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=cf5fe3c71c5a
4 - https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=5dbb19b16ac4
5 - https://lpc.events/event/18/contributions/1937/
6 - https://github.com/bpfverif/agni/actions/runs/17103791870?pr=80
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>