workflows: Cover stable kernels
Let's cover the stable kernels (at least the LTS). Most verifier fixes
are backported there, so it's possible (although not super likely) that
we could end up with a regression on a stable kernel that isn't found in
the bpf or bpf-next trees.

Stable kernels are in the same repository, but under different branches,
so we need to update the tree configuration parameters to be
subsys_tree_branch and default to master for the branch if unspecified.

In addition, for stable kernels, we can select a more precise kernel
version than the latest when running Agni. This commit also updates the
latest kernel version to v6.17 for the bpf and bpf-next trees; it
doesn't make any different right now.

Finally, the torvalds/linux tree is removed because we're unlikely to
find any regression there. Changes from bpf-next end up there, so any
regression in torvalds/linux would be found in bpf/bpf-next first.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>