workflows: Condition modular mode on insns for older kernels
Commit [1] upstream fixed an issue for BPF_{XOR,AND,OR} and their 32 bit
variants when running Agni in modular mode. That commit was merged in
v6.10. Thus, we can't run Agni in modular mode against those ops on
previous kernels.

The present commit therefore conditions modular mode to the instructions
for kernels v6.1, v5.15, v5.10, and v5.4. We also need to generate the
BPF_SYNC encodings when not running in modular mode.

1 - https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=1f586614f3ff
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>