workflows: Run 64 and 32bits insns in same job
There's a hard limit of 256 jobs per workflow on GitHub. With the new
stable trees, we are over that limit. This commit changes the workflow
to run the 32 and 64bits variants of each op in the same job, thus
almost halving the number of jobs.

There shouldn't be any impact on runtime because we're currently limited
by BPF_SYNC, which doesn't have a 32bits counterpart.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>