llvm-to-smt, workflows: Stop relying on BPF_SYNC
The current assumption of modular mode is that BPF_SYNC will be verified
separately from other operations. With the introduction of
BPF_SYNC{1,2,3} in [1], it's now possible to verify those three
operations instead of BPF_SYNC.

Hence, there is no reason to block modular mode on kernels <= 5.19.
Older kernels down to 4.14 can support modular mode.

This commit also removes BPF_SYNC from the end-to-end tests, as it's
currently by far the slowest-to-verify operation and, as described
above, isn't really needed to verify the range analysis.

1 - 84b3ea68ed03 ("Verify reg_bounds_sync faster with further divide-and-conquer")
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>