llvm-to-smt: Use Coccinnelle for sync removal
This commit replaces the manual text parsing in Python in
reg_bounds_sync_calls_remove() by a Coccinelle script. Coccinelle is
already used in the kernel and was designed specifically for this sort
of source-to-source C transformations. The intention is to simplify and
harden the wrappers. This commit starts with the
reg_bounds_sync_calls_remove() but other wrapper logic can be replaced
in the future.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>