llvm-to-smt, workflows: Require v5.4.210
The previous commit requires v5.4.210 or older to be able to run Agni.
We therefore can't run on v5.4.0 anymore. This commit makes that
explicit in the version check and updates the workflow.

Note we need to switch to the stable tree, because Linus' tree doesn't
have the patch releases.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>