llvm-to-smt: remove same bpf_reg_state* comparison from verifier.c
Commit d43ad9da8 [1] upstream causes llvm-to-smt to fail, because it
causes clang to emit code that checks whether two pointers are equal.

The corresponding C code snippet is from is_scalar_branch_taken() in
verifier.c. This function accepts two bpf_reg_state pointers. The
upstream commit d43ad9da8 adds a check to see if they are equal.

static int is_scalar_branch_taken(struct bpf_reg_state *reg1, struct bpf_reg_state *reg2,
				  u8 opcode, bool is_jmp32) {
    ...

    if (reg1 == reg2) {
        ...
        // exit early with return 0 or 1
    }
}

The commit was introduced to address the case where the jump condition
actually compares the register to itself, e.g.  if r0 > r0.

This causes clang to emit the following code pattern:

%cmp1 = icmp eq %struct.bpf_reg_state* %reg1, %reg2

, an icmp on pointers, not something that llvm-to-smt can support. In
Agni, we assume the inputs to the verifier functions (e.g. dst_reg and
src_reg pointers) are disjoint and never alias. So a check like dst_reg
== src_reg is not meaningful in the context of our memory model.

Agni is meant for detecting logic errors in the updating of the
bpf_reg_state bounds, that happens only when the dst_reg and src_reg
pointers are distinct.

All of that to say, we can safely "undo" the upstream commit and ensure
that llvm-to-smt continues to work.

This commit preprocesses the verifier.c code to undo the upstream
commit.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/bpf/bpf-next.git/commit/?id=d43ad9da8052eda714ca

Signed-off-by: Harishankar Vishwanathan <harishankar.vishwanathan@gmail.com>